api_app/requirements.txt
# API
azure-core==1.26.1
aiohttp==3.8.3
azure-cosmos==4.3.0
azure-identity==1.12.0
azure-mgmt-cosmosdb==8.0.0
azure-mgmt-compute==29.1.0
azure-mgmt-costmanagement==3.0.0
azure-storage-blob==12.14.1
azure-servicebus==7.8.1
azure-eventgrid==4.9.1
fastapi[all]==0.89.1
fastapi-utils==0.2.1
gunicorn==20.1.0
jsonschema[format_nongpl]==4.17.1
msal==1.21.0
opencensus-ext-azure==1.1.7
opencensus-ext-logging==0.1.1
PyJWT==2.6.0
uvicorn[standard]==0.20.0
semantic-version==2.10.0
pytz~=2022.7
python-dateutil~=2.8.2
azure-mgmt-resource==22.0.0
pandas==1.5.2
msrestazure
